add a xib
