+++ WMF Framework/BatchEditToolbarViewController.xib
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="13771" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina4_7" orientation="portrait">
+        <adaptation id="fullscreen"/>
+    </device>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="13772"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="BatchEditToolbarViewController" customModule="WMF">
+            <connections>
+                <outlet property="separatorView" destination="fWi-hH-Agp" id="fQ2-sX-qf2"/>
+                <outlet property="stackView" destination="gnJ-Br-g7q" id="57k-GK-oCc"/>
+                <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" insetsLayoutMarginsFromSafeArea="NO" id="i5M-Pr-FkT">
+            <rect key="frame" x="0.0" y="0.0" width="374" height="69"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="fWi-hH-Agp" userLabel="Separator View">
+                    <rect key="frame" x="0.0" y="0.0" width="374" height="0.5"/>
+                    <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                    <constraints>
+                        <constraint firstAttribute="height" constant="0.5" id="cb5-Ls-nDk"/>
+                    </constraints>
+                </view>
+                <stackView opaque="NO" contentMode="scaleToFill" distribution="fillEqually" alignment="center" translatesAutoresizingMaskIntoConstraints="NO" id="gnJ-Br-g7q">
+                    <rect key="frame" x="0.0" y="0.0" width="374" height="69"/>
+                </stackView>
+            </subviews>
+            <constraints>
+                <constraint firstItem="gnJ-Br-g7q" firstAttribute="trailing" secondItem="i5M-Pr-FkT" secondAttribute="trailing" id="Cqp-jL-o01"/>
+                <constraint firstItem="fWi-hH-Agp" firstAttribute="leading" secondItem="i5M-Pr-FkT" secondAttribute="leading" id="I1D-Ye-H4m"/>
+                <constraint firstAttribute="trailing" secondItem="fWi-hH-Agp" secondAttribute="trailing" id="UTe-Jg-03m"/>
+                <constraint firstItem="gnJ-Br-g7q" firstAttribute="leading" secondItem="i5M-Pr-FkT" secondAttribute="leading" id="VPd-za-J4z"/>
+                <constraint firstItem="fWi-hH-Agp" firstAttribute="width" secondItem="i5M-Pr-FkT" secondAttribute="width" id="ksQ-qM-8FW"/>
+                <constraint firstItem="fWi-hH-Agp" firstAttribute="top" secondItem="i5M-Pr-FkT" secondAttribute="top" id="lhX-uf-Re3"/>
+                <constraint firstItem="gnJ-Br-g7q" firstAttribute="bottom" secondItem="i5M-Pr-FkT" secondAttribute="bottom" id="nGQ-lm-NsB"/>
+                <constraint firstItem="gnJ-Br-g7q" firstAttribute="top" secondItem="i5M-Pr-FkT" secondAttribute="top" id="oBT-Ak-lBs"/>
+            </constraints>
+            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
+            <point key="canvasLocation" x="354" y="-264.5"/>
+        </view>
+    </objects>
+</document>
